Ask for the hosts upgrade after the end
The tool is curently asking to raise the cluster and DC compatibility
version after the managers update, but it may not be possible if the
hosts are not up to date as well. This patch adds the request for the
hypervisors update.

Change-Id: I1d387b9dd2a1ba4537adf771498dde553303dfe6
Bug-Url: https://bugzilla.redhat.com/1697964
Signed-off-by: Roman Hodain <rhodain@redhat.com>
(cherry picked from commit dd3d79a1c648a7a4942d56045325b18d8bb12e04)